Finalize plugin setup
