+++ .gitignore
@@ -1,4 +1,4 @@
 __pycache__/
-logs/narrative_trace.jsonl
+logs/
 reports/
 .venv/
