+++ main.py
@@ -1,36 +1,14 @@
 from fastapi import FastAPI
 from fastapi.responses import FileResponse
-from exo_fin_gpt.core.backtesting import run_backtest
+from exo_fin_gpt.exo_interface_api.exo_interface_api import router
 
 app = FastAPI()
+app.include_router(router)
 
 
 @app.get('/health')
 def health():
     return {'status': 'ok'}
-
-
-@app.post('/predict')
-def predict(prices: list[float]):
-    metrics = run_backtest(prices)
-    return metrics
-
-
-@app.get('/risk')
-def risk():
-    # Placeholder risk metric
-    return {'risk': 'low'}
-
-
-@app.get('/evaluate')
-def evaluate():
-    # Return last metrics from report if available
-    try:
-        with open('reports/predictivity_report.md') as f:
-            content = f.read()
-    except FileNotFoundError:
-        content = 'No report available'
-    return {'report': content}
 
 
 @app.get('/openapi.yaml', include_in_schema=False)
